pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0", "versioningit"]
build-backend = "setuptools.build_meta"

[project]
name = "dandischema"
description = "Schemata for DANDI archive project"
readme = {file = "README.md", content-type = "text/markdown"}
license = "Apache-2.0"
authors = [
    {name = "DANDI developers", email = "team@dandiarchive.org"},
]
maintainers = [
    {name = "Yaroslav O. Halchenko", email = "debian@onerussian.com"},
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.10"
dependencies = [
    "jsonschema[format]",
    "packaging>=14.0",
    "pydantic-settings",
    "pydantic[email]~=2.4",
    "requests",
    "zarr_checksum",
]
dynamic = ["version"]

[project.urls]
Homepage = "http://dandiarchive.org"
"Source Code" = "https://github.com/dandi/dandischema"

[project.optional-dependencies]
style = [
    "flake8",
    "pre-commit",
]
test = [
    "anys",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-rerunfailures",
]
all = [
    "dandischema[style]",
    "dandischema[test]",
]

[tool.hatch.envs.default]
installer = "uv"
python = "3.10"

# Env for auto converting `dandischema.models` to LinkML schema and back to Pydantic models
# To invoke conversion scripts provided by the environment, execute the following respective commands:
#   `hatch run linkml-auto-converted:2linkml`    # Convert dandischema.models -> dandischema/models.yaml
#   `hatch run linkml-auto-converted:2pydantic`  # Convert dandischema/models.yaml -> dandischema/models_linkml.py
[tool.hatch.envs.linkml-auto-converted]
dependencies = [
    "black",  # This allows `gen-pydantic` to format the generated Pydantic models with Black formatting
    "dandi",  # DANDI client, used by scripts that fetch dandiset metadata from the archive
    "linkml==1.10.0",
    "linkml-runtime==1.10.0",
    "pydantic2linkml @ git+https://github.com/dandi/pydantic2linkml.git",
    "ruamel.yaml"
]
[tool.hatch.envs.linkml-auto-converted.scripts]
2linkml = "pydantic2linkml -l INFO -M dandischema/models_merge.yaml -O dandischema/models_overlay.yaml dandischema.models | sed -e 's,0x[0-9a-ef]*,0xADDRESS,g' -E -e 's,([^[:space:]]+):[0-9]+,\\1:NUMBER,g' | tools/linkml_conversion_tools/sanitize-yaml > dandischema/models.yaml"
2pydantic = "gen-pydantic --black dandischema/models.yaml | sed -E -e 's,[a-z]+COLON,,g' > dandischema/models_linkml.py"
2json = "d=dandischema/models_linkml; rm -rf $d && mkdir -p $d && for t in Dandiset Asset PublishedDandiset PublishedAsset; do gen-json-schema -t $t dandischema/models.yaml >| $d/$t.json; done"
pydantic2json = "d=dandischema/models_pydantic; rm -rf $d && mkdir -p $d && python tools/pubschemata.py $d && mv $d/*/*json $d/"

[tool.setuptools.packages.find]
namespaces = true
include = ["dandischema*"]

[tool.versioningit]
default-version = "0+unknown"

[tool.versioningit.vcs]
exclude = ["schema-*"]

[tool.versioningit.format]
# Same format as versioneer
distance = "{base_version}+{distance}.{vcs}{rev}"
dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"
distance-dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"

[tool.versioningit.write]
file = "dandischema/_version.py"

[tool.black]
exclude = '\.eggs|\.git|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist'

[tool.isort]
profile = "black"
force_sort_within_sections = true
reverse_relative = true
known_first_party = ["dandischema"]

[tool.mypy]
allow_incomplete_defs = false
allow_untyped_defs = false
ignore_missing_imports = false
no_implicit_optional = true
implicit_reexport = false
local_partial_types = true
pretty = true
show_error_codes = true
show_traceback = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
plugins = ["pydantic.mypy"]
exclude = ['^dandischema/_version\.py$']

[[tool.mypy.overrides]]
module = "dandischema._version"
follow_imports = "skip"

[tool.pydantic-mypy]
init_forbid_extra = true
